Language Service - Named Entity Recognition (NER)/requirements.txt

azure.ai.textanalytics
